requirements.txt
SQLAlchemy==2.0.12
uvicorn==0.22.0
fastapi==0.95.1
pydantic==1.10.7
psycopg2-binary==2.9.6
requests==2.30.0
Jinja2==3.1.2
python-multipart==0.0.6